Abstract states data
